+++ app/assets/javascripts/3dbio_viewer/src/domain/entities/Fragment2.ts
@@ -158,6 +158,24 @@ export interface Interval {
     end: number;
 }
 
+type FragmentP = Fragment | Fragment2;
+
+export function isCovered(alignment: Interval[], fragment: FragmentP): boolean {
+    return alignment.some(
+        interval => fragment.start >= interval.start && fragment.end <= interval.end
+    );
+}
+
+export function isPartiallyCovered(alignment: Interval[], fragment: FragmentP): boolean {
+    return alignment.some(
+        interval => fragment.start <= interval.end && fragment.end >= interval.start
+    );
+}
+
+export function isNotCovered(alignment: Interval[], fragment: FragmentP): boolean {
+    return !isCovered(alignment, fragment) && !isPartiallyCovered(alignment, fragment);
+}
+
 export function getIntervalKey<T extends LooseInterval>(obj: T): string {
     return [obj.start, obj.end].join("-");
 }
